[rocm_regression] Resolve review comments
