+++ test/CMakeLists.txt
@@ -137,6 +137,7 @@ if(BUILD_TESTS)
     # Add rccl-UnitTestsFixtures binary
     list(APPEND RCCL_TEST_EXECUTABLES rccl-UnitTestsFixtures)
 
+
     set(TEST_FIXTURE_SOURCE_FILES
       AltRsmiTests.cpp
       AllocTests.cpp
@@ -180,4 +181,3 @@ if(BUILD_TESTS)
   endforeach()
 
 endif()
-
